test: move chat component tests to message-ui
